introduce "/gps" and "/gps list" commands as aliases for "/video" and "/videos"
This commit introduces a new command, "/gps," serving as a more native and convenient alias for the existing "/video" command and "/gps list" command as an alias for "/videos". Both commands are added to the help window.
Furthermore, some changes have been made to the help window, and a slight correction has been applied to the docstring at the very beginning of the file.